requirements.txt
django==3.2.18
djangorestframework==3.12
asgiref==3.3.2
pip==23.0
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.6
wheel==0.37.1
Pillow==9.3.0
django-taggit==1.2.0
psycopg2==2.9.5
factory-boy==3.2.1
easy-thumbnails==2.8.1

# TODO: Add development packages here
flake8
coverage
ipdb
pre-commit
django-debug-toolbar
django-extensions
django-coverage-plugin
